requirements.txt
torch>=1.10
torchvision>=0.11
ftfy
regex
tqdm
pillow
scikit-learn
streamlit
git+https://github.com/openai/CLIP.git
